Agent config generation from agentspeak code
* Add translator script to root path
* Add header file generation based on agentspeak code from data dir
* Fix intention suspension
