--- config/python/agent.json
@@ -1,57 +1,0 @@
-{
-    "belief":[
-       {
-          "statement":"statement_1",
-          "state":"false"
-       },
-       {
-          "statement":"statement_2",
-          "state":"false"
-       }
-    ],
-    "event":[
-       {
-          "event_operator":"BELIEF_ADDITION",
-          "statement":"statement_1"
-       },
-       {
-          "event_operator":"BELIEF_DELETION",
-          "statement":"duper_event"
-       }
-    ],
-    "plan":[
-       {
-          "event_operator":"BELIEF_ADDITION",
-          "statement":"statement_1",
-          "context":[
-             {
-                "context_condition":[
-                   {
-                      "statement":"statement_1",
-                      "is_true":"true"
-                   },
-                   {
-                      "statement":"statement_2",
-                      "is_true":"false"
-                   }
-                ]
-             }
-          ],
-          "body":[
-             {
-                "instruction":[
-                   {
-                      "statement":"add_goal",
-                      "operator":"GOAL_ADDITION",
-                      "body_type":"GOAL"
-                   },
-                   {
-                      "statement":"do_action",
-                      "body_type":"ACTION"
-                   }
-                ]
-             }
-          ]
-       }
-    ]
- }
