--- config/python/functions.h
@@ -1,14 +1,0 @@
-bool update_statement_1(bool var)
-{
-  return true;
-}
-
-bool action_test_1(bool var)
-{
-  return true;
-}
-
-bool action_do_action(bool var)
-{
-  return false;
-}
