+++ lib/parser/src/translator/EventOperatorType.java
