+++ src/agent_loop.cpp
@@ -9,21 +9,12 @@
 #include "../lib/bdi/event_base.h"
 #include "../lib/bdi/plan_base.h"
 #include "../lib/bdi/intention_base.h"
-#include "example/configuration.h"
+#include "config/configuration.h"
 
-//#include <iostream>
 #include <unistd.h>
-
-//bool compile_test = true;
 
 int main()
 {
-//  int bb_size = 4;
-//  int eb_size = 15;
-//  int pb_size = 4;
-//  int ib_queue_size = 4;
-//  int ib_stack_size = 2;
-
   AgentSettings agent;
 
   BeliefBase * beliefs = agent.get_belief_base();
@@ -42,11 +33,7 @@ int main()
     // Checks if there are events to be processed
     if (!events->is_empty())
     {
-//      std::cout << "Is full: " << events->is_full() << std::endl;
       event_to_process = events->get_event();
-//      std::cout << event_to_process->get_statement().get_name() << std::endl;
-//      std::cout << "Is empty: " << events->is_empty() << std::endl;
-//      std::cout << "--------" << std::endl;
       if(event_to_process)
       {
         plan_to_act = plans->revise(event_to_process, beliefs);
@@ -68,11 +55,6 @@ int main()
     }
 
     sleep(1);
-
-//    if (compile_test == true)
-//    {
-//      break;
-//    }
   }
 
   return 0;
