--- src/example/functions.h
@@ -1,54 +1,0 @@
-#ifndef FUNCTIONS_H_
-#define FUNCTIONS_H_
-
-#include <iostream>
-
-bool action_action_1(bool var)
-{
-  std::cout << "Action 1" << std::endl;
-  return true;
-}
-
-bool action_action_2(bool var)
-{
-  std::cout << "Action 2" << std::endl;
-  return true;
-}
-
-bool action_action_3(bool var)
-{
-  std::cout << "Action 3" << std::endl;
-  return false;
-}
-
-bool action_action_4(bool var)
-{
-  std::cout << "Action 4" << std::endl;
-  return true;
-}
-
-bool action_action_5(bool var)
-{
-  std::cout << "Action 5" << std::endl;
-  return true;
-}
-
-bool action_action_6(bool var)
-{
-  std::cout << "Action 6" << std::endl;
-  return true;
-}
-
-bool action_action_7(bool var)
-{
-  std::cout << "Action 7" << std::endl;
-  return false;
-}
-
-bool action_action_8(bool var)
-{
-  std::cout << "Action 8" << std::endl;
-  return true;
-}
-
-#endif /* FUNCTIONS_H_ */
